+++ src/components/shared/slider-input.ts
@@ -28,10 +28,10 @@ class SliderInput extends HTMLElement {
     const range = this.querySelector<HTMLInputElement>("input[type=range]")!;
     const number = this.querySelector<HTMLInputElement>("input[type=number]")!;
 
-    range.value = number.value = this.value || this.getAttribute("value") || "50";
     range.min = number.min = this.getAttribute("min") || "0";
     range.max = number.max = this.getAttribute("max") || "100";
     range.step = number.step = this.getAttribute("step") || "1";
+    range.value = number.value = this.value || this.getAttribute("value") || "50";
 
     range.addEventListener("input", this.handleEvent.bind(this));
     number.addEventListener("input", this.handleEvent.bind(this));